fix(B-464): split review model and schema
