+++ backend/app/models/__init__.py
@@ -5,7 +5,7 @@ from app.models.conversation_scenario import (
     ConversationScenario,
     ConversationScenarioStatus,
 )
-from app.models.review import Review, ReviewCreate, ReviewRead
+from app.models.review import Review
 from app.models.scenario_preparation import (
     ScenarioPreparation,
     ScenarioPreparationCreate,
@@ -81,6 +81,4 @@ __all__ = [
     'ConfigType',
     'AdminDashboardStats',
     'Review',
-    'ReviewCreate',
-    'ReviewRead',
 ]
